ext/ckanext-entraid_authenticator/requirements.txt
msal==1.28.1
python-dotenv